Refactor DNSHandler — split resolution, takeover detection, and categorisation (#43)
Extracts TakeoverDetector (dangling CNAME, NS takeover, evidence) and
DomainCategoriser (static pattern matching) from DNSHandler so each class
has a single clear responsibility. DNSHandler retains resolution-only logic
and delegates takeover checks via a composed TakeoverDetector.

All 138 tests pass.
